// apps/portal/app/enums/itemActionsEnum.ts
export enum ItemActionsEnum {
  share = 'share',
  shareSlack = 'shareSlack',
  shareWhatsApp = 'shareWhatsApp',
  download = 'download',
  move = 'move',
  delete = 'delete',
  createJiraIssue = 'createJiraIssue',
  createTrelloIssue = 'createTrelloIssue',
  moveToWorkspace = 'moveToWorkspace',
  editPermissions = 'editPermissions',
  restore = 'restore',
}